+++ packages/amplify-cli/src/commands/console.ts
@@ -3,18 +3,6 @@ import { stateManager, open } from 'amplify-cli-core';
 
 const providerName = 'awscloudformation';
 
-const getAdminUIName = async context => {
-  let adminUiName = 'Amplify admin UI';
-  const promises = context.runtime.plugins.map(async plugin => {
-    if (plugin.commands.includes('overrideAdminUiName')) {
-      adminUiName = await context.amplify.invokePluginMethod(context, plugin.name, undefined, 'overrideAdminUiName', [context]);
-    }
-  });
-
-  await Promise.all(promises);
-  return adminUiName;
-};
-
 export const run = async context => {
   let consoleUrl = getDefaultURL();
 
@@ -32,7 +20,6 @@ export const run = async context => {
     const { envName } = localEnvInfo;
     const { Region, AmplifyAppId } = teamProviderInfo[envName][providerName];
 
-    let adminUiName = await getAdminUIName(context);
     if (envName && AmplifyAppId) {
       consoleUrl = constructStatusURL(Region, AmplifyAppId, envName);
       const providerPlugin = await import(context.amplify.getProviderPlugins(context).awscloudformation);
@@ -42,7 +29,7 @@ export const run = async context => {
           name: 'choice',
           message: 'Which site do you want to open?',
           choices: [
-            { name: 'Admin', message: adminUiName },
+            { name: 'Admin', message: 'Amplify Studio' },
             { name: 'Console', message: 'AWS console' },
           ],
         });