+++ packages/amplify-cli/src/execution-manager.ts
@@ -40,32 +40,6 @@ export function isContainersEnabled(context) {
 }
 
 async function selectPluginForExecution(context: Context, pluginCandidates: PluginInfo[]): Promise<PluginInfo> {
-  const pluginCandidatesCategorySet = new Set<string>();
-
-  pluginCandidates.forEach(plugin => {
-    pluginCandidatesCategorySet.add(plugin.manifest.name);
-  });
-
-  // overrided packageName format : @aws-amplify/amplify-category-<catgoryName>
-  const pluginName = pluginCandidatesCategorySet.values().next().value;
-  if (pluginCandidatesCategorySet.size == 1 && overriddenCategories.includes(pluginName)) {
-    if (FeatureFlags.getBoolean(`overrides.${pluginName}`)) {
-      const pluginWithOverrides = pluginCandidates.find(plugin => {
-        return plugin.packageName === `@aws-amplify/amplify-category-${pluginName}`;
-      });
-      if (pluginWithOverrides !== undefined) {
-        return pluginWithOverrides;
-      }
-    } else {
-      const pluginWithOutOverrides = pluginCandidates.find(plugin => {
-        return plugin.packageName === `amplify-category-${pluginName}`;
-      });
-      if (pluginCandidates.length === 2 && pluginWithOutOverrides !== undefined) {
-        return pluginWithOutOverrides;
-      }
-    }
-  }
-
   let result = pluginCandidates[0];
 
   let promptForSelection = true;